requirements.txt
requests
io
tkinter
urllib
pillow
webbrowser